core: remove System.Binary.Itoa and System.Binary.Atoi syscalls
And move their tests to native StdLib.
