+++ pkg/core/interop/interopnames/names.go
@@ -2,13 +2,11 @@ package interopnames
 
 // Names of all used interops.
 const (
-	SystemBinaryAtoi                         = "System.Binary.Atoi"
 	SystemBinaryBase58Decode                 = "System.Binary.Base58Decode"
 	SystemBinaryBase58Encode                 = "System.Binary.Base58Encode"
 	SystemBinaryBase64Decode                 = "System.Binary.Base64Decode"
 	SystemBinaryBase64Encode                 = "System.Binary.Base64Encode"
 	SystemBinaryDeserialize                  = "System.Binary.Deserialize"
-	SystemBinaryItoa                         = "System.Binary.Itoa"
 	SystemBinarySerialize                    = "System.Binary.Serialize"
 	SystemCallbackCreate                     = "System.Callback.Create"
 	SystemCallbackCreateFromMethod           = "System.Callback.CreateFromMethod"
@@ -54,13 +52,11 @@ const (
 )
 
 var names = []string{
-	SystemBinaryAtoi,
 	SystemBinaryBase58Decode,
 	SystemBinaryBase58Encode,
 	SystemBinaryBase64Decode,
 	SystemBinaryBase64Encode,
 	SystemBinaryDeserialize,
-	SystemBinaryItoa,
 	SystemBinarySerialize,
 	SystemCallbackCreate,
 	SystemCallbackCreateFromMethod,
